Handling opting in/out the PC module, Reverting demo app to default settings and adding documentation
