+++ demos/appsflyer-react-native-app/android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.appsflyer.billing2",
+    package = "com.appsflyerexample",
 )
 
 android_resource(
     name = "res",
-    package = "com.appsflyer.billing2",
+    package = "com.appsflyerexample",
     res = "src/main/res",
 )
