+++ demos/appsflyer-react-native-app/android/app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
   xmlns:tools="http://schemas.android.com/tools"
-  package="com.appsflyer.billing2">
+  package="com.appsflyerexample">
 
     <uses-permission android:name="android.permission.INTERNET" />
