+++ demos/appsflyer-react-native-app/android/gradle.properties
@@ -27,4 +27,4 @@ android.enableJetifier=true
 # Version of flipper SDK to use with React Native
 FLIPPER_VERSION=0.93.0
 # Uncomment to enable purchase connector
-appsflyer.enable_purchase_connector=true
+# appsflyer.enable_purchase_connector=true
