+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-native-appsflyer",
-  "version": "6.15.1",
+  "version": "6.15.2",
   "description": "React Native Appsflyer plugin",
   "main": "index.js",
   "types": "index.d.ts",
